+++ src/Phan/Config.php
@@ -424,20 +424,11 @@ class Config
         // to stdout instead of parsing and analyzing files.
         'dump_parsed_file_list' => false,
 
-        // Include a progress bar in the output
+        // Include a progress bar in the output.
         'progress_bar' => false,
 
-        // The probability of actually emitting any progress
-        // bar update. Setting this to something very low
-        // is good for reducing network IO and filling up
-        // your terminal's buffer when running phan on a
-        // remote host.
-        // Set this to 0 to use *only* progress_bar_sample_interval.
-        'progress_bar_sample_rate' => 0.000,
-
         // If this much time (in seconds) has passed since the last update,
-        // then update the progress bar (Ignores progress_bar_sample_rate).
-        // Set this to INF to only use progress_bar_sample_rate.
+        // then update the progress bar.
         'progress_bar_sample_interval' => 0.1,
 
         // The number of processes to fork off during the analysis